Rename adaptive sampling testing.
