+++ src/testing/CMakeLists.txt
@@ -23,7 +23,7 @@ set(TESTS_SOURCE
     taylor_green_vortex_energy_check.cpp
     taylor_green_vortex_restart_check.cpp
     pod_adaptive_sampling.cpp
-    adaptive_sampling_testing.cpp
+    pod_adaptive_sampling_testing.cpp
     reduced_order.cpp
     time_refinement_study.cpp
     time_refinement_study_reference.cpp
